Add GET handler for user mappings
Moved the user mapping helper function to the mappings handler file,
since that makes more sense. Some of the code in the function had to be
taken out because the error response when a user mapping doesn't exist
isn't always the same. This unfortunately makes the code less DRY
overall.